Update README.md
In Opensuse Leap 15.2 the chain would not start syncing until this change was performed. Not so sure whether this is actually needed, but IPs in the configuration file seemed to be old (from testnet-1).
In addition, deleting the file and fetching it new would start with empty persistent_peers, so I assume deleting them should not hurt.
Please review this PR as I'm a bit new to github.
PS: I believe indentation of my change is not fully OK, feel free to change it if needed.